Renamed test_sqlitecachedb_upgrade.py to test_db_upgrader.py
